+++ app/src/main/java/org/breezyweather/sources/china/ChinaService.kt
@@ -199,7 +199,6 @@ class ChinaService @Inject constructor(
         }
     }
 
-
     override fun requestLocationSearch(
         context: Context, query: String
     ): Observable<List<Location>> {